refactor: `coordinate`, `car` 경로 이동 반영
