+++ src/test/java/racingcar/domain/car/dto/input/CreateCarsRacingTest.java
@@ -6,8 +6,8 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.exception.DuplicateCarNameException;
-import racingcar.exception.NoCarsException;
+import racingcar.exception.car.DuplicateCarNameException;
+import racingcar.exception.car.NoCarsException;
 
 final class CreateCarsRacingTest {
